Clean code listItemHandler
