+++ event-planner.html
@@ -79,11 +79,6 @@
             <h2 id="todo-popup-h2">Titel från Js</h2>
           </header>
           <div id="todo-popup-info">
-            <p>
-              Lorem ipsum dolor sit amet consectetur adipisicing elit. Ex, esse? Architecto saepe ab est, accusamus
-              molestias corrupti libero cumque voluptates. Ut in provident quae distinctio veritatis iste. A, aliquam
-              facilis!
-            </p>
           </div>
           <div id="todo-popup-status">
             <ul></ul>
